+++ doc/gallery/cuboid.py
@@ -3,10 +3,11 @@
 import fresnel
 import numpy
 import PIL
+import sys
 
 data = numpy.load('cuboids.npz')
 
-scene = fresnel.Scene(fresnel.Device(mode='cpu'))
+scene = fresnel.Scene()
 scene.lights = fresnel.light.lightbox()
 W, H, D = data['width']
 poly_info = fresnel.util.convex_polyhedron_from_vertices([
@@ -33,9 +34,14 @@ geometry.outline_material = fresnel.material.Material(color=(0.95, 0.93, 0.88),
                                                       roughness=0.1,
                                                       metal=1.0)
 
-scene.camera = fresnel.camera.fit(scene, view='front')
+scene.camera = fresnel.camera.Orthographic.fit(scene, view='front')
 out = fresnel.pathtrace(scene, samples=64, light_samples=32, w=580, h=580)
 PIL.Image.fromarray(out[:], mode='RGBA').save('cuboid.png')
 
-out = fresnel.pathtrace(scene, samples=256, light_samples=16, w=1380, h=1380)
-PIL.Image.fromarray(out[:], mode='RGBA').save('cuboid-hires.png')
+if len(sys.argv) > 1 and sys.argv[1] == 'hires':
+    out = fresnel.pathtrace(scene,
+                            samples=256,
+                            light_samples=16,
+                            w=1380,
+                            h=1380)
+    PIL.Image.fromarray(out[:], mode='RGBA').save('cuboid-hires.png')